feat(gen): serialization-aware value-object accessors (Theme.of chain on web)

+++ ewt.api/src/main/java/dev/equo/ewt/Chip.java
@@ -45,63 +45,129 @@ public class Chip extends StatelessWidget implements ChipI {
     return ChipChipBuilder.chipChip();
   }
   public Widget avatar() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipAvatar not supported on web");
     return new Widget(ChipObjSt.avatar(st)) {};
   }
   public Widget label() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipLabel not supported on web");
     return new Widget(ChipObjSt.label(st)) {};
   }
   public TextStyle labelStyle() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipLabelStyle", getId());
+      java.lang.foreign.MemorySegment __st = dev.equo.ewt.ffm.TextStyleObjSt.allocate(__s.arena);
+      dev.equo.ewt.ffm.TextStyleObjSt.id(__st, __nid);
+      return new TextStyle(__st);
+    }
     return new TextStyle(ChipObjSt.labelStyle(st));
   }
   public EdgeInsetsGeometry labelPadding() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipLabelPadding not supported on web");
     return new EdgeInsetsGeometry(ChipObjSt.labelPadding(st)) {};
   }
   public BorderSide side() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipSide", getId());
+      java.lang.foreign.MemorySegment __st = dev.equo.ewt.ffm.BorderSideObjSt.allocate(__s.arena);
+      dev.equo.ewt.ffm.BorderSideObjSt.id(__st, __nid);
+      return new BorderSide(__st);
+    }
     return new BorderSide(ChipObjSt.side(st));
   }
   public OutlinedBorder shape() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipShape not supported on web");
     return new OutlinedBorder(ChipObjSt.shape(st)) {};
   }
   public Clip clipBehavior() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipClipBehavior not supported on web");
     return Clip.values()[ChipObjSt.clipBehavior(st)];
   }
   public boolean autofocus() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipAutofocus not supported on web");
     return intToBool(ChipObjSt.autofocus(st));
   }
   public Color backgroundColor() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipBackgroundColor", getId());
+      return new Color(__nid);
+    }
     return new Color(ChipObjSt.backgroundColor(st));
   }
   public EdgeInsetsGeometry padding() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipPadding not supported on web");
     return new EdgeInsetsGeometry(ChipObjSt.padding(st)) {};
   }
   public VisualDensity visualDensity() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipVisualDensity", getId());
+      java.lang.foreign.MemorySegment __st = dev.equo.ewt.ffm.VisualDensityObjSt.allocate(__s.arena);
+      dev.equo.ewt.ffm.VisualDensityObjSt.id(__st, __nid);
+      return new VisualDensity(__st);
+    }
     return new VisualDensity(ChipObjSt.visualDensity(st));
   }
   public Widget deleteIcon() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipDeleteIcon not supported on web");
     return new Widget(ChipObjSt.deleteIcon(st)) {};
   }
   public Color deleteIconColor() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipDeleteIconColor", getId());
+      return new Color(__nid);
+    }
     return new Color(ChipObjSt.deleteIconColor(st));
   }
   public String deleteButtonTooltipMessage() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipDeleteButtonTooltipMessage not supported on web");
     return ChipObjSt.deleteButtonTooltipMessage(st).getString(0);
   }
   public MaterialTapTargetSize materialTapTargetSize() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipMaterialTapTargetSize not supported on web");
     return MaterialTapTargetSize.values()[ChipObjSt.materialTapTargetSize(st)];
   }
   public double elevation() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) throw new UnsupportedOperationException("chipElevation not supported on web");
     return ChipObjSt.elevation(st);
   }
   public Color shadowColor() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipShadowColor", getId());
+      return new Color(__nid);
+    }
     return new Color(ChipObjSt.shadowColor(st));
   }
   public Color surfaceTintColor() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipSurfaceTintColor", getId());
+      return new Color(__nid);
+    }
     return new Color(ChipObjSt.surfaceTintColor(st));
   }
   public BoxConstraints avatarBoxConstraints() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipAvatarBoxConstraints", getId());
+      java.lang.foreign.MemorySegment __st = dev.equo.ewt.ffm.BoxConstraintsObjSt.allocate(__s.arena);
+      dev.equo.ewt.ffm.BoxConstraintsObjSt.id(__st, __nid);
+      return new BoxConstraints(__st);
+    }
     return new BoxConstraints(ChipObjSt.avatarBoxConstraints(st));
   }
   public BoxConstraints deleteIconBoxConstraints() {
+    if (dev.equo.ewt.web.EwtWebTransport.isWebMode()) {
+      SerializingWidgetConstructors __s = (SerializingWidgetConstructors) factories;
+      int __nid = __s.recordAccessor("chipDeleteIconBoxConstraints", getId());
+      java.lang.foreign.MemorySegment __st = dev.equo.ewt.ffm.BoxConstraintsObjSt.allocate(__s.arena);
+      dev.equo.ewt.ffm.BoxConstraintsObjSt.id(__st, __nid);
+      return new BoxConstraints(__st);
+    }
     return new BoxConstraints(ChipObjSt.deleteIconBoxConstraints(st));
   }
   @Override
